chore: add test comments

+++ test/Raffl/entries/Raffl.EntriesWithNative.t.sol
@@ -1,18 +1,9 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.24;
 
-import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import { IERC721 } from "@openzeppelin/contracts/token/ERC721/IERC721.sol";
-import { Initializable } from "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
-import { VRFV2PlusClient } from "@chainlink/contracts/src/v0.8/vrf/dev/libraries/VRFV2PlusClient.sol";
-
 import { Raffl } from "../../../src/Raffl.sol";
 import { Common } from "../../utils/Common.sol";
-import { IRaffl } from "../../../src/interfaces/IRaffl.sol";
-import { RafflFactory } from "../../../src/RafflFactory.sol";
-import { IFeeManager } from "../../../src/interfaces/IFeeManager.sol";
 import { RafflErrors } from "../../../src/libraries/Errors.sol";
-import { VRFCoordinatorV2PlusMock } from "../../mocks/VRFCoordinatorV2PlusMock.sol";
 
 contract RafflEntriesWithNativeTest is Common {
     Raffl raffl;
@@ -47,7 +38,7 @@ contract RafflEntriesWithNativeTest is Common {
         raffl.buyEntries{ value: 123 }(5);
     }
 
-    /// @dev should validate the value when buying entries
+    /// @dev should emit `EntriesBought` event when buying entries
     function test_EmitEntriesBought() public {
         uint256 quantity = 5;
         uint256 value = entryPrice * quantity;
@@ -84,15 +75,13 @@ contract RafflEntriesWithNativeTest is Common {
         uint256 quantity = 7;
         makeUserBuyEntries(raffl, userC, quantity);
 
-        for (uint i = 0; i < quantity; i++) {
+        for (uint256 i = 0; i < quantity; i++) {
             uint256 entryNumber = lastEntry + i;
             address entryOwner = raffl.entriesMap(entryNumber);
             assertEq(entryOwner, userC);
-            
         }
 
         uint256 userTotalEntries = raffl.userEntriesMap(userC);
         assertEq(userTotalEntries, quantity);
-
     }
 }
